CLN: move casey trees tests (#13)
