--- lambdas/casey_trees/complete_data.txt
@@ -1,116 +1,0 @@
-{'@context': 'http://schema.org',
- '@type': 'Event',
- 'description': '&lt;p&gt;Learn how you can protect D.C.\\u2019s trees at this fun, interactive workshop. Stand Up for Trees will prepare you to navigate D.C.\\u2019s unique political geography to help increase tree canopy in your neighborhood and across the District. By the end of the class you will be on your way to reviewing development plans, writing persuasive [&hellip;]&lt;/p&gt;\\n',
- 'endDate': '2019-01-12T15:00:00-05:00',
- 'image': 'https://caseytrees.org/wp-content/uploads/2018/06/stand-up.jpg',
- 'location': {'@type': 'Place',
-              'address': {'@type': 'PostalAddress',
-                          'addressCountry': 'United States',
-                          'addressLocality': 'Washington',
-                          'addressRegion': 'DC',
-                          'postalCode': '20017',
-                          'streetAddress': '3030 12th Street Northeast'},
-              'description': '',
-              'geo': {'@type': 'GeoCoordinates',
-                      'latitude': 38.9284202,
-                      'longitude': -76.99097},
-              'name': 'Casey Trees',
-              'sameAs': '',
-              'telephone': '',
-              'url': 'https://caseytrees.org/venue/casey-trees/'},
- 'name': 'Advocacy in the District: Stand Up for Trees',
- 'offers': {'@type': 'Offer',
-            'price': '0',
-            'url': 'https://caseytrees.org/event/advocacy-in-the-district-stand-up-for-trees-2/'},
- 'organizer': {'@type': 'Person',
-               'description': '',
-               'email': '',
-               'name': 'Casey Trees',
-               'sameAs': 'http://caseytrees.org',
-               'telephone': '',
-               'url': 'https://caseytrees.org/organizer/casey-trees-5/'},
- 'startDate': '2019-01-12T09:45:00-05:00',
- 'url': 'https://caseytrees.org/event/advocacy-in-the-district-stand-up-for-trees-2/'}
-{'@context': 'http://schema.org',
- '@type': 'Event',
- 'description': "&lt;p&gt;Description Join Casey Trees for a conversation about advancing the city\\'s environmental\\u00a0programs\\u00a0in 2019. As we prepare to testify at the D.C. Council\\'s upcoming agency performance hearings, this is your chance to get ideas, ask questions, and collect key information for your testimony. This is an invite-only event for Casey Trees Advocates. After an introductory exercise, [&hellip;]&lt;/p&gt;\\n",
- 'endDate': '2019-01-26T13:00:00-05:00',
- 'image': 'https://caseytrees.org/wp-content/uploads/2018/12/P1010683.jpg',
- 'location': {'@type': 'Place',
-              'address': {'@type': 'PostalAddress',
-                          'addressCountry': 'United States',
-                          'addressLocality': 'Washington',
-                          'addressRegion': 'DC',
-                          'postalCode': '20002',
-                          'streetAddress': '1344 H Street NE'},
-              'description': '',
-              'geo': {'@type': 'GeoCoordinates',
-                      'latitude': 38.900434,
-                      'longitude': -76.9866727},
-              'name': 'Sospeso',
-              'sameAs': '',
-              'telephone': '',
-              'url': 'https://caseytrees.org/venue/sospeso/'},
- 'name': 'Tree Advocates Brunch: Performance Hearing Testimony Workshop',
- 'offers': {'@type': 'Offer',
-            'price': 'Donation',
-            'url': 'https://caseytrees.org/event/tree-advocates-brunch-performance-hearing-testimony-workshop/'},
- 'organizer': {'@type': 'Person',
-               'description': '',
-               'email': '',
-               'name': 'Casey Trees',
-               'sameAs': 'http://caseytrees.org',
-               'telephone': '',
-               'url': 'https://caseytrees.org/organizer/casey-trees-5/'},
- 'startDate': '2019-01-26T10:00:00-05:00',
- 'url': 'https://caseytrees.org/event/tree-advocates-brunch-performance-hearing-testimony-workshop/'}
-{'@context': 'http://schema.org',
- '@type': 'Event',
- 'description': '&lt;p&gt;Do you want to play a greater role in re-treeing D.C.? We need your help to protect and promote trees in our urban forest! Trees 101 provides a foundation in tree anatomy, basic tree identification and an overview of how trees function to provide the benefits we enjoy in the urban forest. The class will [&hellip;]&lt;/p&gt;\\n',
- 'endDate': '2019-02-09T15:00:00-05:00',
- 'image': 'https://caseytrees.org/wp-content/uploads/2018/12/a1ed661548bf815eae944860bf897fba.jpg',
- 'location': {'@type': 'Place',
-              'address': {'@type': 'PostalAddress'},
-              'description': '',
-              'name': 'TBD',
-              'sameAs': '',
-              'telephone': '',
-              'url': 'https://caseytrees.org/venue/tbd/'},
- 'name': 'Trees 101',
- 'offers': {'@type': 'Offer',
-            'price': '0',
-            'url': 'https://caseytrees.org/event/trees-101-2/'},
- 'organizer': {'@type': 'Person',
-               'description': '',
-               'email': '',
-               'name': 'Casey Trees',
-               'sameAs': 'http://caseytrees.org',
-               'telephone': '',
-               'url': 'https://caseytrees.org/organizer/casey-trees-5/'},
- 'startDate': '2019-02-09T10:30:00-05:00',
- 'url': 'https://caseytrees.org/event/trees-101-2/'}
-{'@context': 'http://schema.org',
- '@type': 'Event',
- 'description': '&lt;p&gt;Please note: This is an invite-only event for Casey Trees Advocates. Volunteers must be able to attend both Part 1: The Discussion and Part 2: The Workshop in order to participate in this tree advocates meeting series. To register, select tickets for both events.\\u00a0 Part 1: The Discussion Tuesday, March 5, 2019 6:30-8 pm Join\\u00a0the [&hellip;]&lt;/p&gt;\\n',
- 'endDate': '2019-03-05T20:00:00-05:00',
- 'image': 'https://caseytrees.org/wp-content/uploads/2018/12/P1010719.jpg',
- 'location': {'@type': 'Place',
-              'address': {'@type': 'PostalAddress'},
-              'description': '',
-              'name': 'TBD',
-              'sameAs': '',
-              'telephone': '',
-              'url': 'https://caseytrees.org/venue/tbd/'},
- 'name': 'Tree Advocates Meetings Parts 1 &#038; 2: Budget Hearings Discussion &#038; Workshop',
- 'offers': {'@type': 'Offer',
-            'price': 'Donation',
-            'url': 'https://caseytrees.org/event/tree-advocates-meetings-parts-1-2-budget-hearings-discussion-workshop/'},
- 'organizer': {'@type': 'Person',
-               'description': '',
-               'email': '',
-               'name': 'Casey Trees',
-               'sameAs': 'http://caseytrees.org',
-               'telephone': '',
-               'url': 'https://caseytrees.org/organizer/casey-trees-5/'},
- 'startDate': '2019-03-05T18:30:00-05:00',
- 'url': 'https://caseytrees.org/event/tree-advocates-meetings-parts-1-2-budget-hearings-discussion-workshop/'}
