--- lambdas/casey_trees/output.txt
@@ -1,60 +1,0 @@
-{'Event Cost': '0',
-  'Event Description': '&lt;p&gt;Learn how you can protect D.C.\\u2019s trees at this fun, interactive workshop. Stand Up for Trees will prepare you to navigate D.C.\\u2019s unique political geography to help increase tree canopy in your neighborhood and across the District. By the end of the class you will be on your way to reviewing development plans, writing persuasive [&hellip;]&lt;/p&gt;\\n',
-  'Event End Date': '2019-01-12',
-  'Event End Time': '15:00',
-  'Event Featured Image': 'https://caseytrees.org/wp-content/uploads/2018/06/stand-up.jpg',
-  'Event Name': 'Advocacy in the District: Stand Up for Trees',
-  'Event Organizer Name(s) or ID(s)': 'Casey Trees',
-  'Event Start Date': '2019-01-12',
-  'Event Start Time': '09:45',
-  'Event Time Zone': 'America/New_York',
-  'Event Venue Name': 'Casey Trees',
-  'Event Website': 'https://caseytrees.org/event/advocacy-in-the-district-stand-up-for-trees-2/',
-  'address': 'United States Washington DC 3030 12th Street Northeast 20017 PostalAddress',
-  'latitude': 38.9284202,
-  'longitude': -76.99097},
- {'Event Cost': 'Donation',
-  'Event Description': "&lt;p&gt;Description Join Casey Trees for a conversation about advancing the city\\'s environmental\\u00a0programs\\u00a0in 2019. As we prepare to testify at the D.C. Council\\'s upcoming agency performance hearings, this is your chance to get ideas, ask questions, and collect key information for your testimony. This is an invite-only event for Casey Trees Advocates. After an introductory exercise, [&hellip;]&lt;/p&gt;\\n",
-  'Event End Date': '2019-01-26',
-  'Event End Time': '13:00',
-  'Event Featured Image': 'https://caseytrees.org/wp-content/uploads/2018/12/P1010683.jpg',
-  'Event Name': 'Tree Advocates Brunch: Performance Hearing Testimony Workshop',
-  'Event Organizer Name(s) or ID(s)': 'Casey Trees',
-  'Event Start Date': '2019-01-26',
-  'Event Start Time': '10:00',
-  'Event Time Zone': 'America/New_York',
-  'Event Venue Name': 'Sospeso',
-  'Event Website': 'https://caseytrees.org/event/tree-advocates-brunch-performance-hearing-testimony-workshop/',
-  'address': 'United States Washington DC 1344 H Street NE 20002 PostalAddress',
-  'latitude': 38.900434,
-  'longitude': -76.9866727},
- {'Event Cost': '0',
-  'Event Description': '&lt;p&gt;Do you want to play a greater role in re-treeing D.C.? We need your help to protect and promote trees in our urban forest! Trees 101 provides a foundation in tree anatomy, basic tree identification and an overview of how trees function to provide the benefits we enjoy in the urban forest. The class will [&hellip;]&lt;/p&gt;\\n',
-  'Event End Date': '2019-02-09',
-  'Event End Time': '15:00',
-  'Event Featured Image': 'https://caseytrees.org/wp-content/uploads/2018/12/a1ed661548bf815eae944860bf897fba.jpg',
-  'Event Name': 'Trees 101',
-  'Event Organizer Name(s) or ID(s)': 'Casey Trees',
-  'Event Start Date': '2019-02-09',
-  'Event Start Time': '10:30',
-  'Event Time Zone': 'America/New_York',
-  'Event Venue Name': 'TBD',
-  'Event Website': 'https://caseytrees.org/event/trees-101-2/',
-  'address': 'PostalAddress',
-  'latitude': 'no geo',
-  'longitude': 'no geo'},
- {'Event Cost': 'Donation',
-  'Event Description': '&lt;p&gt;Please note: This is an invite-only event for Casey Trees Advocates. Volunteers must be able to attend both Part 1: The Discussion and Part 2: The Workshop in order to participate in this tree advocates meeting series. To register, select tickets for both events.\\u00a0 Part 1: The Discussion Tuesday, March 5, 2019 6:30-8 pm Join\\u00a0the [&hellip;]&lt;/p&gt;\\n',
-  'Event End Date': '2019-03-05',
-  'Event End Time': '20:00',
-  'Event Featured Image': 'https://caseytrees.org/wp-content/uploads/2018/12/P1010719.jpg',
-  'Event Name': 'Tree Advocates Meetings Parts 1 &#038; 2: Budget Hearings Discussion &#038; Workshop',
-  'Event Organizer Name(s) or ID(s)': 'Casey Trees',
-  'Event Start Date': '2019-03-05',
-  'Event Start Time': '18:30',
-  'Event Time Zone': 'America/New_York',
-  'Event Venue Name': 'TBD',
-  'Event Website': 'https://caseytrees.org/event/tree-advocates-meetings-parts-1-2-budget-hearings-discussion-workshop/',
-  'address': 'PostalAddress',
-  'latitude': 'no geo',
-  'longitude': 'no geo'}
